% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToNumeric.R
\name{convertToNumeric}
\alias{convertToNumeric}
\title{Convert the value of genotype into 0,1,2}
\usage{
convertToNumeric(snp_data)
}
\arguments{
\item{snp_data}{The snp_data matrix.}
}
\value{
The genodata with the value of 0,1,2.
}
\description{
This function converts the value of genotype into 0,1,2 form 0/0,0/1,1/1.
}
\examples{
convertTonumeric(snp_data)
}
