% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_from_snp.R
\name{get_gene_from_snp}
\alias{get_gene_from_snp}
\title{Get the candidate genes in regions based on significant SNPs of GWAS results.}
\usage{
get_gene_from_snp(gff, sig.snp, distance = 75000, file.save = TRUE,
  file.type = "csv", gff.chrom = NA, snp.chrom = NA, geneid = NA,
  pvalue = NA, gene_start = NA, gene_end = NA, snp_location = NA,
  verbose = TRUE, ...)
}
\arguments{
\item{gff}{a data frame of all the gene (transcript), must have column names.}

\item{sig.snp}{a data frame of significant SNPs.}

\item{distance}{numeric (bp), it is to define the region. The default is 50000, you need to choose it based on the LD distance in your study.}

\item{file.save}{a logical, if file.output=TRUE, the result will be saved.
if file.output=FALSE, the result will be printed. The default is TRUE.}

\item{file.type}{a character, users can choose the different output formats,
so far, "csv", "txt", "xlsx" can be selected by users. The default is "csv".}

\item{gff.chrom}{Name of the column containing the chromosome identifers in the gff file; default is NA.}

\item{snp.chrom}{Name of the column containing the chromosome identifers in the snp.sig file; default is NA.}

\item{geneid}{Name of the column containing the geneid in gff file; default is NA.}

\item{pvalue}{Name of the column containing the p values in snp.sig file; default is NA.}

\item{snp_location}{Name of the column containing the snp position in snp.sig file; default is NA.}
}
\value{
a data.frame contain the candidate genes with start,end,genid etc
}
\description{
This function is developed to get the candidate genes in regions based on significant SNPs of GWAS results.
}
\examples{
get_gene_from_snp(gff,sig.snp)
}
\author{
Tao Yan <\email{tyan@zju.edu.cn}> |
<\href{https://taoyan.netlify.com/}{https://taoyan.netlify.com/}>
}
