% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsnpInfo.R
\name{getsnpInfo}
\alias{getsnpInfo}
\title{Obtain the snp info from the vcf file}
\usage{
getsnpInfo(file)
}
\arguments{
\item{file}{The vcf file name.}
}
\value{
The snp info (ID, POS).
}
\description{
This function defined to get the snp info (ID, POS) from the vcf file.
}
\examples{
snpInfo(sample.vcf)
}
