% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmanhattan.R
\name{ggmanhattan}
\alias{ggmanhattan}
\title{Make manhattan plot with full ggplot customizability}
\usage{
ggmanhattan(gwasres, snp = NA, bp = NA, chrom = NA, pvalue = NA,
  index = NA, file.output = FALSE, file = "png", output = "Trait",
  dpi = 300, vlinetype = "solid", vlinesize = 0.75,
  title = "Manhattan Plot", color = c("#FF8C00", "#556B2F"),
  pointsize = 1.25, verbose = TRUE, ...)
}
\arguments{
\item{gwasres}{a data frame of gwas results.}

\item{snp}{Name of the column containing SNP identifiers; default is NA.}

\item{bp}{Name of the column containing the SNP positions; default is NA.}

\item{chrom}{Name of the column containing the chromosome identifers; default is NA.}

\item{pvalue}{Name of the column containing the p values; default is NA.}

\item{file.output}{a logical, if file.output=TRUE, the result will be saved.
if file.output=FALSE, the result will be printed. The default is TRUE.}

\item{file}{a character, users can choose the different output formats of plot,
so far, "jpeg", "pdf", "png", "tiff" can be selected by users. The default is "png".}

\item{output}{a character, the name of your trait. The default is "Trait"}

\item{dpi}{a number, the picture element for .jpeg, .png and .tiff files. The default is 300.}

\item{vlinetype}{the type of vline (geom_vline()). The default is "solid".}

\item{vlinesize}{the size of the vline. The default is 0.75.}

\item{title}{the title of manhattan plot. The default is "Manhattan Plot".}

\item{color}{the colors of alternate chromosome. The default is "#FF8C00" and "#556B2F"}

\item{pointsize}{the size of point. The default is 1.25.}
}
\value{
a manhattan plot based on ggplot2.
}
\description{
This function is provided to make manhattan plot with full ggplot customizability. So next
we can customize the manhattan plot with kinds of functions of ggplot2 and add additional layers.
}
\examples{
ggmanhattan(gwas_test)
}
\author{
Tao Yan <\email{tyan@zju.edu.cn}> |
<\href{https://taoyan.netlify.com/}{https://taoyan.netlify.com/}>
}
