% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_country.R
\name{hello_country}
\alias{hello_country}
\title{hello_country}
\usage{
hello_country(countries)
}
\arguments{
\item{countries}{string vertices countaining countries.}
}
\description{
hello_country
}
\examples{
hello_country(countries)
}
