% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyLDheatMap.R
\name{MyLDheatMap}
\alias{MyLDheatMap}
\title{Get the LDhaetmap from the vcf file (plink format)}
\usage{
MyLDheatMap(vcffile)
}
\arguments{
\item{vcffile}{The plink format vcf file.}
}
\value{
the LDheatmap.
}
\description{
This function defined to obtain the LDheatmap from the vcf file directly.
}
\examples{
MyLDheatMap(test_vcf)
}
