% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpMat.R
\name{snpMat}
\alias{snpMat}
\title{Get the snpMatrix from the vcf file (plink format)}
\usage{
snpMat(vcffile)
}
\arguments{
\item{vcffile}{The vcf file (plink format).}
}
\value{
The snpMAtrix for the visulization of LDheatmap.
}
\description{
This function defined to obtain the snpMatrix(with snp ID) from the vcf file
}
\examples{
snpMat(vcffile)
}
