% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tt.R
\name{theme_tt}
\alias{theme_tt}
\title{Add Triptease theme to a ggplot}
\usage{
theme_tt(font = "sans")
}
\arguments{
\item{font}{The font to be used by titles and axes - Use Roboto Medium}
}
\description{
Add Triptease theme to a ggplot
}
\examples{
ggplot2::ggplot(iris, ggplot2::aes(Sepal.Length, Sepal.Width)) + ggplot2::geom_point() + theme_tt()
}
