% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt_colors.R
\docType{data}
\name{tt_colors}
\alias{tt_colors}
\title{Use the Triptease Brand colors.}
\format{An object of class \code{character} of length 6.}
\usage{
tt_colors
}
\value{
the TT color palette.
}
\description{
Use the Triptease Brand colors.
}
\examples{
ggplot2::ggplot(iris, ggplot2::aes(Sepal.Length, Sepal.Width, color = Species)) +
ggplot2::geom_point() +
ggplot2::scale_color_manual(values = tt_colors)
}
\keyword{datasets}
