% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyLDheatMap.R
\name{MyLDheatMap}
\alias{MyLDheatMap}
\title{Get the LDhaetmap from the vcf file (plink format) directly}
\usage{
MyLDheatMap(vcffile, file.output = TRUE, file = "png",
  output = "region", title = "region:", verbose = TRUE, dpi = 300)
}
\arguments{
\item{vcffile}{The plink format vcf file. More detail can see View(test_vcf)}

\item{file.output}{a logical, if file.output=TRUE, the result will be saved.
if file.output=FALSE, the result will be printed. The default is TRUE}

\item{file}{a character, users can choose the different output formats of plot, so far, "jpeg", "pdf", "png", "tiff" can be selected by users. The default is "png".}

\item{title}{a character, the title of the LDheatmap will be "The LDheatmap of title".
the default is "region:". I suggest users use your own title.}

\item{verbose}{whether to print the reminder.}

\item{dpi}{a number, the picture element for .jpeg, .png and .tiff files. The default is 300.}
}
\value{
the LDheatmap.
}
\description{
This function defined to obtain the LDheatmap from the vcf file directly.
}
\examples{
MyLDheatMap(system.file("extdata","test.vcf", package = "ttplot"), title="your title")
}
\author{
Tao Yan <\email{tyan@zju.edu.cn}> |
<\href{https://taoyan.netlify.com/}{https://taoyan.netlify.com/}>
}
